# backend/openapi/app.yaml
...
"openapi": "3.1.0"
"info":
  "title": "Vigía — interfaz de la plataforma"
  "version": "0.1.0"
"paths":
...
  "/zones/{zone_id}/coverage":
    "get":
      "tags":
        - "cobertura"
      "summary": "Línea de tiempo de cobertura de la zona en [from, to), tope de 31 días"
      "operationId": "timeline_zones__zone_id__coverage_get"
      "parameters":
        - "name": "zone_id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
            "title": "Zone Id"
        - "name": "from"
          "in": "query"
          "required": true
          "schema":
            "type": "string"
            "title": "From"
        - "name": "to"
          "in": "query"
          "required": true
          "schema":
            "type": "string"
            "title": "To"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/CoverageTimelineOut"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
  "/zones/{zone_id}/coverage/at":
    "get":
      "tags":
        - "cobertura"
      "summary": "Estado compuesto de cobertura de la zona en un instante"
      "operationId": "status_at_zones__zone_id__coverage_at_get"
      "parameters":
        - "name": "zone_id"
          "in": "path"
          "required": true
          "schema":
            "type": "string"
            "format": "uuid"
            "title": "Zone Id"
        - "name": "instant"
          "in": "query"
          "required": true
          "schema":
            "type": "string"
            "title": "Instant"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/CoverageStatusOut"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
  "/integrity/verify":
    "post":
      "tags":
        - "integridad"
      "summary": "Pide verificar una cadena a demanda; la verificación corre en el worker"
      "operationId": "verify_integrity_verify_post"
      "requestBody":
        "content":
          "application/json":
            "schema":
              "$ref": "#/components/schemas/VerifyRequest"
        "required": true
      "responses":
        "202":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/VerificationAccepted"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
  "/integrity/results":
    "get":
      "tags":
        - "integridad"
      "summary": "Último resultado de la verificación de cada cadena a su alcance"
      "operationId": "results_integrity_results_get"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/IntegrityResultsOut"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
  "/integrity/checkpoints":
    "get":
      "tags":
        - "integridad"
      "summary": "Último punto de control firmado de cada cadena a su alcance"
      "operationId": "checkpoints_integrity_checkpoints_get"
      "responses":
        "200":
          "description": "Successful Response"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/CheckpointsOut"
        "default":
          "description": "Error genérico (ApiError)"
          "content":
            "application/json":
              "schema":
                "$ref": "#/components/schemas/ApiErrorBody"
...
"components":
  "schemas":
...
    "ApiErrorBody":
      "additionalProperties": false
      "description": "Cuerpo JSON de toda respuesta de error (domain-entities §4.4, pendiente nº 33)."
      "properties":
        "code":
          "$ref": "#/components/schemas/ApiErrorCode"
        "detail_code":
          "anyOf":
            - "maxLength": 64
              "pattern": "[a-z][a-z0-9]*(?:_[a-z0-9]+)+"
              "type": "string"
            - "type": "null"
          "default": null
          "title": "Detail Code"
        "message_es":
          "maxLength": 256
          "minLength": 1
          "title": "Message Es"
          "type": "string"
        "correlation_id":
          "format": "uuid"
          "title": "Correlation Id"
          "type": "string"
        "retry_after_seconds":
          "anyOf":
            - "maximum": 3600
              "minimum": 1
              "type": "integer"
            - "type": "null"
          "default": null
          "title": "Retry After Seconds"
      "required":
        - "code"
        - "message_es"
        - "correlation_id"
      "title": "ApiErrorBody"
      "type": "object"
    "ApiErrorCode":
      "description": "``api_error_code`` (domain-entities §1 con el pendiente nº 33): lista cerrada de U-02."
      "enum":
        - "invalid_request"
        - "not_found"
        - "unauthenticated"
        - "second_factor_required"
        - "forbidden"
        - "throttled"
        - "rate_limited"
        - "payload_too_large"
        - "conflict"
        - "temporarily_unavailable"
        - "internal_error"
        - "privacy_notice_required"
        - "period_too_long"
        - "zone_without_node"
        - "storage_unavailable"
      "title": "ApiErrorCode"
      "type": "string"
...
    "ChainOut":
      "properties":
        "kind":
          "type": "string"
          "enum":
            - "ledger"
            - "audit"
          "title": "Kind"
        "plant_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Plant Id"
      "additionalProperties": false
      "type": "object"
      "required":
        - "kind"
        - "plant_id"
      "title": "ChainOut"
...
    "CheckpointOut":
      "properties":
        "chain":
          "$ref": "#/components/schemas/ChainOut"
        "sequence":
          "type": "integer"
          "title": "Sequence"
        "entry_id":
          "type": "string"
          "format": "uuid"
          "title": "Entry Id"
        "entry_hash":
          "type": "string"
          "title": "Entry Hash"
        "covered_sequence":
          "type": "integer"
          "title": "Covered Sequence"
        "covered_hash":
          "type": "string"
          "title": "Covered Hash"
        "taken_at":
          "type": "string"
          "title": "Taken At"
        "key_id":
          "type": "string"
          "title": "Key Id"
        "signature":
          "type": "string"
          "title": "Signature"
      "additionalProperties": false
      "type": "object"
      "required":
        - "chain"
        - "sequence"
        - "entry_id"
        - "entry_hash"
        - "covered_sequence"
        - "covered_hash"
        - "taken_at"
        - "key_id"
        - "signature"
      "title": "CheckpointOut"
    "CheckpointsOut":
      "properties":
        "organization_id":
          "type": "string"
          "format": "uuid"
          "title": "Organization Id"
        "checkpoints":
          "items":
            "$ref": "#/components/schemas/CheckpointOut"
          "type": "array"
          "title": "Checkpoints"
      "additionalProperties": false
      "type": "object"
      "required":
        - "organization_id"
        - "checkpoints"
      "title": "CheckpointsOut"
...
    "CoverageStatusOut":
      "properties":
        "zone_id":
          "type": "string"
          "format": "uuid"
          "title": "Zone Id"
        "instant":
          "type": "string"
          "title": "Instant"
        "state":
          "type": "string"
          "title": "State"
        "causes":
          "items":
            "type": "string"
          "type": "array"
          "title": "Causes"
        "layer":
          "type": "string"
          "title": "Layer"
        "clock_basis":
          "type": "string"
          "title": "Clock Basis"
        "source_record_ids":
          "items":
            "type": "string"
            "format": "uuid"
          "type": "array"
          "title": "Source Record Ids"
      "additionalProperties": false
      "type": "object"
      "required":
        - "zone_id"
        - "instant"
        - "state"
        - "causes"
        - "layer"
        - "clock_basis"
        - "source_record_ids"
      "title": "CoverageStatusOut"
    "CoverageTimelineOut":
      "properties":
        "organization_id":
          "type": "string"
          "format": "uuid"
          "title": "Organization Id"
        "plant_id":
          "type": "string"
          "format": "uuid"
          "title": "Plant Id"
        "zone_id":
          "type": "string"
          "format": "uuid"
          "title": "Zone Id"
        "period":
          "additionalProperties":
            "type": "string"
          "type": "object"
          "title": "Period"
        "intervals":
          "items":
            "$ref": "#/components/schemas/IntervalOut"
          "type": "array"
          "title": "Intervals"
        "node_intervals":
          "items":
            "$ref": "#/components/schemas/IntervalOut"
          "type": "array"
          "title": "Node Intervals"
        "summary":
          "$ref": "#/components/schemas/SummaryOut"
      "additionalProperties": false
      "type": "object"
      "required":
        - "organization_id"
        - "plant_id"
        - "zone_id"
        - "period"
        - "intervals"
        - "node_intervals"
        - "summary"
      "title": "CoverageTimelineOut"
...
    "IntegrityResultsOut":
      "properties":
        "results":
          "items":
            "$ref": "#/components/schemas/ResultOut"
          "type": "array"
          "title": "Results"
      "additionalProperties": false
      "type": "object"
      "required":
        - "results"
      "title": "IntegrityResultsOut"
    "IntervalOut":
      "properties":
        "starts_at":
          "type": "string"
          "title": "Starts At"
        "ends_at":
          "type": "string"
          "title": "Ends At"
        "duration_ms":
          "type": "integer"
          "title": "Duration Ms"
        "layer":
          "type": "string"
          "title": "Layer"
        "state":
          "type": "string"
          "title": "State"
        "causes":
          "items":
            "type": "string"
          "type": "array"
          "title": "Causes"
        "subject":
          "anyOf":
            - "$ref": "#/components/schemas/SubjectOut"
            - "type": "null"
        "source_record_ids":
          "items":
            "type": "string"
            "format": "uuid"
          "type": "array"
          "title": "Source Record Ids"
        "clock_basis":
          "type": "string"
          "title": "Clock Basis"
        "clock_offset_ms":
          "anyOf":
            - "type": "integer"
            - "type": "null"
          "title": "Clock Offset Ms"
      "additionalProperties": false
      "type": "object"
      "required":
        - "starts_at"
        - "ends_at"
        - "duration_ms"
        - "layer"
        - "state"
        - "causes"
        - "subject"
        - "source_record_ids"
        - "clock_basis"
      "title": "IntervalOut"
      "description": "``CoverageInterval`` (domain-entities §3.8)."
...
    "ResultOut":
      "properties":
        "chain":
          "$ref": "#/components/schemas/ChainOut"
        "mode":
          "type": "string"
          "title": "Mode"
        "result":
          "type": "string"
          "title": "Result"
        "from_sequence":
          "type": "integer"
          "title": "From Sequence"
        "to_sequence":
          "type": "integer"
          "title": "To Sequence"
        "head_sequence":
          "type": "integer"
          "title": "Head Sequence"
        "broken_sequence":
          "anyOf":
            - "type": "integer"
            - "type": "null"
          "title": "Broken Sequence"
        "broken_entry_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Broken Entry Id"
        "reason":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Reason"
        "verified_at":
          "anyOf":
            - "type": "string"
            - "type": "null"
          "title": "Verified At"
      "additionalProperties": false
      "type": "object"
      "required":
        - "chain"
        - "mode"
        - "result"
        - "from_sequence"
        - "to_sequence"
        - "head_sequence"
        - "broken_sequence"
        - "broken_entry_id"
        - "reason"
        - "verified_at"
      "title": "ResultOut"
...
    "SubjectOut":
      "properties":
        "kind":
          "type": "string"
          "title": "Kind"
        "camera_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Camera Id"
        "signal_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Signal Id"
      "additionalProperties": false
      "type": "object"
      "required":
        - "kind"
      "title": "SubjectOut"
    "SummaryOut":
      "properties":
        "observable_ms":
          "type": "integer"
          "title": "Observable Ms"
        "degraded_ms":
          "type": "integer"
          "title": "Degraded Ms"
        "not_observable_ms":
          "type": "integer"
          "title": "Not Observable Ms"
        "no_communication_ms":
          "type": "integer"
          "title": "No Communication Ms"
        "never_reported_ms":
          "type": "integer"
          "title": "Never Reported Ms"
        "zone_not_active_ms":
          "type": "integer"
          "title": "Zone Not Active Ms"
      "additionalProperties": false
      "type": "object"
      "required":
        - "observable_ms"
        - "degraded_ms"
        - "not_observable_ms"
        - "no_communication_ms"
        - "never_reported_ms"
        - "zone_not_active_ms"
      "title": "SummaryOut"
...
    "VerificationAccepted":
      "properties":
        "request_id":
          "type": "string"
          "format": "uuid"
          "title": "Request Id"
        "chain":
          "$ref": "#/components/schemas/ChainOut"
        "status":
          "type": "string"
          "const": "queued"
          "title": "Status"
        "requested_at":
          "type": "string"
          "title": "Requested At"
      "additionalProperties": false
      "type": "object"
      "required":
        - "request_id"
        - "chain"
        - "status"
        - "requested_at"
      "title": "VerificationAccepted"
...
    "VerifyRequest":
      "properties":
        "kind":
          "type": "string"
          "enum":
            - "ledger"
            - "audit"
          "title": "Kind"
        "plant_id":
          "anyOf":
            - "type": "string"
              "format": "uuid"
            - "type": "null"
          "title": "Plant Id"
      "additionalProperties": false
      "type": "object"
      "required":
        - "kind"
      "title": "VerifyRequest"
      "description": "La cadena a verificar: ``ledger`` con ``plant_id`` (planta) o sin él (organización), o\n``audit`` (sin planta)."